requirements.txt
Flask==3.0.2
jieba
lxml
fake_useragent
pandas==2.0.3
PyMySQL==1.1.0
snownlp==0.12.3
SQLAlchemy==2.0.21
redis
cryptography
paddlenlp==2.4.2
paddlepaddle==2.4.2
